fix: webui accessibility/contrast fixes, training bug fixes, cleanup stale docs
WebUI (design review fixes):
- Critical: add role="meter" with ARIA attributes to EvalBar
- Major: darken promoted piece color for AA contrast on board
- Major: add aria-labels to PieceTray hand pieces
- Major: add WebSocket reconnection banner with connection state store
- Major: replace fragile JS-derived thumbnail width with CSS clamp()
- Major: pause commentary rotation on hover/focus, respect reduced-motion
- Minor: CJK font fallback, section header size bump, phase badge icons,
  notation toggle aria-label, matchup matrix legend 50% label,
  collapse button hover state, responsive clamp for zoom safety

Training fixes:
- Cast learner_log_probs/values to float in rollout scratch buffers
- Make transformer positional buffers non-persistent (checkpoint compat)
- Set training_enabled flag on role change in opponent_store
- Fix match_type using stale is_trainable vs fresh is_train_now
- Log when trainable match has no rollout collected

Also: remove stale generated bug report docs, add profile_training script.

+++ keisei/training/katago_loop.py
@@ -1139,8 +1139,8 @@ class KataGoTrainingLoop:
                         _scratch_log_probs.zero_()
                         _scratch_values.zero_()
                         if li.numel() > 0:
-                            _scratch_log_probs[li] = sm_result.learner_log_probs
-                            _scratch_values[li] = sm_result.learner_values
+                            _scratch_log_probs[li] = sm_result.learner_log_probs.float()
+                            _scratch_values[li] = sm_result.learner_values.float()
 
                         # Update spectator-facing value estimates so snapshots
                         # reflect current policy predictions (not stale zeros).
